backend/requirements.txt
django
djangorestframework
django-filter
django-cors-headers
djangorestframework-jsonapi